title screen!

+++ include/2d.h
@@ -1,5 +1,5 @@
 #ifndef DRAW2D_H
 #define DRAW2D_H
 #include "things.h"
-void DrawTex(f32, f32, unsigned int, unsigned int, GXTexObj);
+void DrawTex(unsigned int, unsigned int, unsigned int, unsigned int, GXTexObj);
 #endif
